Validate target YAML C flags
