+++ docs/COMPLETION_AUDIT.md
@@ -4138,10 +4138,13 @@ and do not assume Phase 4 is ready without evidence.
   Covered by `emit_json_hir_rejects_hand_authored_json_before_ir_override`.
 - Hand-authored target YAML remains outside the compiler-owned IR path:
   `emit-json target-yaml` validates a minimal schema plus the current C backend
-  schema into `zen.target.v0` JSON, rejects attempts to override
-  compiler-owned type layouts, and rejects unsupported backend code generators.
+  schema into `zen.target.v0` JSON, accepts optional `backend.c_flags`, rejects
+  empty C flag entries, rejects attempts to override compiler-owned type
+  layouts, and rejects unsupported backend code generators.
   Covered by `emit_json_target_yaml_validates_minimal_target_schema`,
   `emit_json_target_yaml_validates_backend_schema`,
+  `emit_json_target_yaml_validates_c_backend_flags`,
+  `emit_json_target_yaml_rejects_empty_c_backend_flags`,
   `emit_json_target_yaml_rejects_layout_overrides`, and
   `emit_json_target_yaml_rejects_unsupported_backend_codegen`.
 - Typechecker boolean literal identifiers parse `true` and `false` through
