+++ tests/docs_truth/phase_audit.rs
@@ -45,6 +45,8 @@ fn phase_plan_records_recovered_progress_and_next_slice() {
         "emit_json_mir_outputs_checked_minimal_function_graph",
         "emit_json_target_yaml_validates_minimal_target_schema",
         "emit_json_target_yaml_validates_backend_schema",
+        "emit_json_target_yaml_validates_c_backend_flags",
+        "emit_json_target_yaml_rejects_empty_c_backend_flags",
         "emit_json_target_yaml_rejects_layout_overrides",
         "emit_json_target_yaml_rejects_unsupported_backend_codegen",
         "build_graph_rejects_unknown_target_dependencies",
