+++ tests/docs_truth/v1_spec.rs
@@ -122,6 +122,8 @@ fn v1_spec_records_phase_one_feature_gates_and_test_backlog() {
         "emit_json_layout_rejects_hand_authored_json_before_layout_override",
         "emit_json_target_yaml_validates_minimal_target_schema",
         "emit_json_target_yaml_validates_backend_schema",
+        "emit_json_target_yaml_validates_c_backend_flags",
+        "emit_json_target_yaml_rejects_empty_c_backend_flags",
         "emit_json_target_yaml_rejects_layout_overrides",
         "emit_json_target_yaml_rejects_unsupported_backend_codegen",
         "build.zen",
